+++ packages/target-electron/runtime-electron/runtime.ts
@@ -215,6 +215,9 @@ class ElectronRuntime implements Runtime {
   }> {
     return ipcBackend.invoke('webxdc.getWebxdcDiskUsage', accountId)
   }
+  async writeClipboardToTempFile(_name: string | undefined): Promise<string> {
+    return ipcBackend.invoke('app.writeClipboardToTempFile')
+  }
   writeTempFileFromBase64(name: string, content: string): Promise<string> {
     return ipcBackend.invoke('app.writeTempFileFromBase64', name, content)
   }